🧪 Add unit tests for ensure_default_tenant and refactor schema tests
- Add comprehensive unit tests for `ensure_default_tenant` in `backend/core/schema.py`.
- Refactor schema-related tests from `tests/test_core_infrastructure.py` to `tests/test_core_schema.py`.
- Implement detailed data verification using mocks for `uuid` and `datetime`.
- Add tests for idempotency, connection configuration, and error handling.
